Move PeersHandleProvider + types to new module reth_network_api::test_utils
