+++ crates/net/network-api/Cargo.toml
@@ -33,6 +33,7 @@ thiserror.workspace = true
 serde = { workspace = true, features = ["derive"], optional = true }
 tokio = { workspace = true, features = ["sync"] }
 auto_impl.workspace = true
+derive_more.workspace = true
 
 [features]
 default = ["serde"]
