+++ crates/net/network-api/src/test_utils/mod.rs
@@ -1,0 +1,5 @@
+//! API for integration testing network components.
+
+pub mod peers_manager;
+
+pub use peers_manager::{PeerCommand, PeersHandle, PeersHandleProvider};
