+++ crates/net/network-types/Cargo.toml
@@ -16,9 +16,6 @@ workspace = true
 reth-network-peers.workspace = true
 reth-net-banlist.workspace = true
 reth-ethereum-forks.workspace = true
-
-# async
-tokio = { workspace = true, features = ["sync"] }
 reth-network-p2p.workspace = true
 
 # io
@@ -28,7 +25,6 @@ serde_json = { workspace = true }
 
 # misc 
 tracing.workspace = true
-derive_more.workspace = true
 
 [features]
 serde = ["dep:serde", "dep:humantime-serde", "reth-network-p2p/serde"]
