+++ crates/net/network-types/src/peers/mod.rs
@@ -1,13 +1,10 @@
 pub mod addr;
-pub mod cmd;
 pub mod config;
-pub mod handle;
 pub mod kind;
 pub mod state;
 
 pub use reth_network_p2p::reputation;
 
-pub use cmd::PeerCommand;
 pub use config::{ConnectionsConfig, PeersConfig};
 pub use reputation::ReputationChangeWeights;
