+++ crates/net/network/src/test_utils/testnet.rs
@@ -12,9 +12,11 @@ use futures::{FutureExt, StreamExt};
 use pin_project::pin_project;
 use reth_chainspec::MAINNET;
 use reth_eth_wire::{protocol::Protocol, DisconnectReason, HelloMessageWithProtocols};
-use reth_network_api::{NetworkInfo, Peers, PeersHandleProvider};
+use reth_network_api::{
+    test_utils::{PeersHandle, PeersHandleProvider},
+    NetworkInfo, Peers,
+};
 use reth_network_peers::PeerId;
-use reth_network_types::PeersHandle;
 use reth_provider::test_utils::NoopProvider;
 use reth_storage_api::{BlockReader, BlockReaderIdExt, HeaderProvider, StateProviderFactory};
 use reth_tasks::TokioTaskExecutor;
