+++ examples/custom-rlpx-subprotocol/src/main.rs
@@ -15,9 +15,9 @@ use std::net::{Ipv4Addr, SocketAddr, SocketAddrV4};
 use reth::builder::NodeHandle;
 use reth_network::{
     config::SecretKey, protocol::IntoRlpxSubProtocol, NetworkConfig, NetworkManager,
-    NetworkProtocols, PeersHandleProvider,
+    NetworkProtocols,
 };
-use reth_network_api::NetworkInfo;
+use reth_network_api::{test_utils::PeersHandleProvider, NetworkInfo};
 use reth_node_ethereum::EthereumNode;
 use reth_provider::test_utils::NoopProvider;
 use subprotocol::{
